Refs #6297 - workaround for incomplete pulp task from sync plan
Also DRYify some code in presenter
